+++ src/client/redux/actions/application.ts
@@ -10,46 +10,51 @@ import {
   SET_TOTAL_PAGES,
 } from "../actionTypes";
 
-import { Job } from "../../types";
+import { ApplicationAction, Job } from "../../types";
 
-export const setCurrentJobs = (currentJobs: Job[]) => ({
+export const setCurrentJobs = (currentJobs: Job[]): ApplicationAction => ({
   type: SET_CURRENT_JOBS,
   payload: { currentJobs },
 });
 
-export const setCurrentPage = (currentPage: number) => ({
+export const setCurrentPage = (currentPage: number): ApplicationAction => ({
   type: SET_CURRENT_PAGE,
   payload: { currentPage },
 });
 
-export const setFullTime = (fullTime: boolean) => ({
+export const setFullTime = (fullTime: boolean): ApplicationAction => ({
   type: SET_FULL_TIME,
   payload: { fullTime },
 });
 
-export const setIsLoading = (isLoading: boolean) => ({
+export const setIsLoading = (isLoading: boolean): ApplicationAction => ({
   type: SET_IS_LOADING,
   payload: { isLoading },
 });
 
-export const setJobs = (jobs: Job[]) => ({ type: SET_JOBS, payload: { jobs } });
+export const setJobs = (jobs: Job[]): ApplicationAction => ({
+  type: SET_JOBS,
+  payload: { jobs },
+});
 
-export const setJobsFetchedAt = (jobsFetchedAt: string) => ({
+export const setJobsFetchedAt = (jobsFetchedAt: string): ApplicationAction => ({
   type: SET_JOBS_FETCHED_AT,
   payload: { jobsFetchedAt },
 });
 
-export const setLocationSearch = (locationSearch: string) => ({
+export const setLocationSearch = (
+  locationSearch: string
+): ApplicationAction => ({
   type: SET_LOCATION_SEARCH,
   payload: { locationSearch },
 });
 
-export const setSearchValue = (searchValue: string) => ({
+export const setSearchValue = (searchValue: string): ApplicationAction => ({
   type: SET_SEARCH_VALUE,
   payload: { searchValue },
 });
 
-export const setTotalPages = (totalPages: number) => ({
+export const setTotalPages = (totalPages: number): ApplicationAction => ({
   type: SET_TOTAL_PAGES,
   payload: { totalPages },
 });